[feature] Add flag for customizing mask size (#11)
Some models, such as qwen and phi-2, has a larger model vocab size than their tokenizer vocab size. The reason for that is the model vocab size will be round up to a power of 2, e.g. 32. In such cases, the mask size should be align with the model vocab size since it is applied to the logits.

This PR adds a parameter mask_vocab_size to GrammarStateMatcher to specify the model vocab size.