[init] GrammarMatcherInitContext and GrammarMatcherInitContextCache (#16)
This PR introduces GrammarMatcherInitContext, which is used to construct
GrammarMatcher fast. A GrammarMatcherInitContext object can be used to
construct multiple GrammarMatcher objects with negligible overhead for
each.

GrammarMatcherInitContextCache is a centralized cache for
GrammarMatcherInitContext. Currently it can cache the init context for
JSON and JSON Schema. It will support more grammar in the future.

This PR also renames GrammarStateMatcher to GrammarMatcher for
simplicity.