[fix] Separate compilation of C++ library and Python binding library (#26)
Our python binding library requires linking to libtorch, which requires setting `-D_GLIBCXX_USE_CXX11_ABI=0` in compilation flags. However, we do not want to set this for C++, as the downstream application linking to it may not set this to be true. So we separate the compilation of C++ lib and python binding lib, and only the latter will set this flag.