[Fix] Avoid directly passing tokenizer to GrammarMatcher (#33)
Now in GrammarMatcher, if we pass a huggingface tokenizer, it will be
converted into xgrammar TokenizerInfo first. This process could be slow.
However, if we can construct TokenizerInfo first and then construct the
GrammarMatcher, the construction of GrammarMatcher would be much faster.

This PR enforces user to construct a TokenizerInfo object first then
pass it into GrammarMatcher, instead passing raw huggingface tokenizer
to it, avoiding the redundant processing.

This PR also adds `<|eos|>` as an EOS token.