[Feature] Optional Token Bitmask (#165)
Sometimes the bitmask is all-true, which means the mask don't need to be applied. This is especially common for function calling cases because we only constrain the part within the tag, and for the part out of tag, all tokens are allowed.


This PR introduces a bool return value for `FillNextTokenBitmask` to represent if the filled bitmask need to applied. If false, user can skip applying the bitmask.

This PR also updates the logic for TagDispatch, so that when the stack top is at an tag dispatch rule (corresponding to the part out of tag), the special tokens are allowed.