Modernize build to support sdists (#190)
Replace the two setup setuptools based build process with
scikit_build_core.build to call cmake.

This adds support for sdists releases of this package, git install
directly of this repo from pip (submodules may make that hard), and
generally simplifies the build process since now all that is required is
doing pip install . in the directory or "python -m build" to build the
wheel and sdist with no other build steps. This also fixes some cross
compiling issues.

This also means that cibuildwheel will work out of the box with github
to build this for more architectures and you won't have to maintain as
much custom github action CI logic.

A subset of tests run in CI as part package building now. 

CUDA is optional again which fixes Mac builds. The optional triton
kernel is also available as a backend but only when the triton package
is also installed in the env.

Updated documentation and install instructions. 

Add attestation support for uploaded wheels.

---------

Co-authored-by: Michał Górny <mgorny@gentoo.org>
Co-authored-by: Yixin Dong <ubospica@gmail.com>