[Feature] Support model vocab size being less than tokenizer (#237)
Some multimodal models have some added tokens, such as <|image|>, that
are not considered in lm_head or generated by the model. In this case,
the model's vocab size (i.e. the dimension of lm_head) is smaller than
the tokenizer. This PR supports this case.

Note that after this PR, the `vocab_size` parameter of `TokenizerInfo`
should always be the model's vocab_size (i.e. lm_head's size).

This PR also refactors the TokenizerInfo class to make the logic
clearer. Now the enum VocabType will be serialized as integer, not
string.