[Style] Remove unused headers (#219)
Currently, there are many headers in C++ files that are not used
directly. We remove all of them (thanks to clangd).

Moreover, it seems that an `inline` is missing in
`cpp/support/thread_pool.h`, which may lead to a potential violation of
ODR in the future.

Maybe this PR should be merged only after a C++ test ci is ready to
ensure correctness.