Fallback to triton if we fail to compile for CUDA (#223)
The CUDA driver might be available but the full CUDA toolkit, ninja, and
gcc/clang might not be installed so we won't have enough tools to
compile the kernel, and in this case use triton as a fallback.

We could consider doing an AOT of the kernel and not JITing the kernel
extension, however this really blows the packaging complexity out of the
water since we would have to target multiple PyTorch extension ABIs.