Add MLX kernel to support Apple Silicon (#287)
As described in
https://github.com/mlc-ai/xgrammar/pull/285#issue-2961934829, there are
four ways to implement apply_token_bitmask for Apple Silicon in MLX.
This PR is approach 4, which is from Angelos Katharopoulos. Kudos to
Angelos!

Also, as described in
https://github.com/mlc-ai/xgrammar/pull/285#issuecomment-2768166617,
this approach is the simplest and the most performant option. After the
merge of this PR, I will close the rest PRs implementing other
approaches:

- https://github.com/mlc-ai/xgrammar/pull/283
- https://github.com/mlc-ai/xgrammar/pull/285

For our reference, I am highlighting the experiment report at
https://github.com/mlc-ai/xgrammar/pull/285#issuecomment-2768166617 with
content copied to the following:

- Manually written Metal GPU kernel compiled with
`mx.fast.metal_kernel`: 0.0002715208800509572 seconds
- Python-code compiled with `mx.compile`: 0.00022429581731557845 seconds
- Naive Python implementation:  0.31162573751062156 seconds