[Feature] Use an earley parser to replace the current pushdown automata (#308)
At present, xgrammar adopts a non-deterministic pushdown automata (NPDA)
in matching. This PR manages to use an Earley Parser, to reduce the time
complexity. Moreover, some optimizations are implemented to reduce the
time to compile grammars.

In this PR:
- NPDA is replaced with Earley Parser.
- right recursions are optimized to speed up.
- "First Character Mask" is implemented to reduce the time to compile
grammars.
- "Almost Recursion Rules" like `rule ::= [a-zA-Z0-9] rule | "\""` will
be optimized to speed up in compiling.
- Uncertain tokens are checked more efficiently in matching.

With the new structure:
- The grammar compiling performance is improved significantly. This PR
is about 600% faster than the current version to compile the built-in
json grammar. Currently, XGrammar needs about 27000 us to compile the
built-in json grammar. In this PR, it takes about 4500 us to compile it.
- Time to fill the next token mask is reduced significantly.
- It can support implicit left recursion rules now.