Fix: fix wrong LRU cache implementation (#352)
1. Fix the bug when an exception happens in creating an entry
2. Fix missing clear map when LRU is not enabled (unlimited case)

When `std::packaged_task` gets an exception inside the task, the
associated `std::shared_future` will rethrow the exception on calling
`get()` method. When this happens in LRU operations, this will lead to
underlying cache inconsistency.

We fix this by catching all the possible exceptions (hopefully) when
`get()` is called.