[Minor] Replace all Result<...>::Ok with ResultOk (#355)
When returning with OK, we don't need to actually specify the type in
most cases. So we implement a wrapper class `PartialResult` to enable
users to write code like: `ResultOK` and `ResultErr`. Still, we have
fallback to specify the result type or the error type, and the code will
look like `ResultOK<T>` and `ResultErr<E>`