Fix apply_bitmask logit for both CPU and triton versions when shape and stride doesn't match (#390)
# Motivation
In vLLM, in order to address [Issue
#19493](https://github.com/vllm-project/vllm/issues/19493), [PR
#19565](https://github.com/vllm-project/vllm/pull/19565) served as a
workaround to mitigate the issue by switching to use torch.compile
version instead of triton version.

# Root cause
With some investigation, we found that the error would happen when
logits.stride()[0] != logits.shape[-1].

# Changes
- Fix CPU and Triton-version apply_grammar_bitmask for this scenario 
  - use stride[0] instead of shape[-1] when access rows
- Add a unit test to reproduce the errors

# Tests
We've verified
- New unit tests would failed on trunk
- New unit tests passed with the change (and no new failure, however,
there're a lot of failing tests in trunk, 47 failed in
tests/python/test_token_bitmask_operations.py)

# Followup
We will follow up with a benchmark comparison to see if we should bring
back triton-version or use the new cuda version on vLLM side.

---------

Signed-off-by: Jialin Ouyang <Jialin.Ouyang@gmail.com>