[Fix]Fix the grammar_compiler. (#395)
This PR fixes a minor bug in Grammar Compiler, which is reported by
#382. The bug is led by a small optimization, which is used to reduce
the time to fill the `tmp_rejected_indices`. In the implementation, jf
`tmp_rejected_indices == false`, then it shouldn't be set to `true`
again, since some infomation will miss. Regretfully, when
`tmp_rejected_indices` is initialized with `false`, and the
`tmp_rejected_indices` is set to `true` until the end of the function,
then the incorrect `tmp_rejected_indices` will be passed.


Signed-off-by: Yuchuan <blemiade_qinchuan@sjtu.edu.cn>

---------

Signed-off-by: Yuchuan <blemiade_qinchuan@sjtu.edu.cn>