[Optim] Optim repetition expressions when the max repetition time is unbounded. (#403)
This PR optims the behavior of repetition expression when the max
repetition time is unbounded. It keeps the invariant of the repetition
states, instead of setting the upper as a large number. This can
significantly speed lower the uncertainty when the max repetition time
is unbounded, and the min repetition time is very small, like this
schema:
```
{
    "type": "string",
    "minLength": 1
}
```

Signed-off-by: Yuchuan
[blemiade_qinchuan@sjtu.edu.cn](mailto:blemiade_qinchuan@sjtu.edu.cn)